Refactor assembly handlers to use Intel syntax with no prefix and add ATT syntax prefix
- Updated all assembly handler functions to use `.intel_syntax noprefix` for consistency.
- Added `.att_syntax prefix` at the end of assembly blocks for compatibility.
- Changed function declarations to use `__attribute__((naked))` for better optimization.
- Modified `CubeModLoader` macros to reflect the new assembly syntax.
- Improved `CubeModLoader` build process with a new Makefile for easier compilation.
- Added PowerShell build script for unified build automation.
- Updated README with new build instructions and options.
- Enhanced item price calculation logic for better performance and accuracy.

+++ CubeModLoader/callbacks/item/cube__Item__OnGetSellingPrice.h
@@ -1,10 +1,12 @@
 #pragma once
+#include <cmath>
+#include <algorithm>
 #include "../../CWSDK/cwsdk.h"
 
 extern "C" int cube__Item__OnGetSellingPrice(cube::Item* item)
 {
 	int buyingPrice = item->GetBuyingPrice();
-	int sellingPrice = std::max<float>(0.5f * std::sqrt(buyingPrice), 1);
+	int sellingPrice = (int)std::max<float>(0.5f * std::sqrt((float)buyingPrice), 1.0f);
 
 	for (uint8_t priority = 0; priority <= 4; priority += 1) {
 		for (DLL* dll : modDLLs) {
